remove invalid TODOs
`typeExpr` produces an expression for the given reflected type, omitting
certain parts of the type fundamentally changes it making it invalid for
whatever the original use was. These were all incorrect TODOs, and were
resolved previously by introducing `RequiresUnexported` which signals
this is the case and the type cannot be created without unexported
access.

In other words, `AST` now handles this logic.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>