detect and handle cyclic data structures
This seems to be a fairly comprehensive approach to detecting cyclic
data structures. The emitted code for now is just the data structure
repeated twice followed by a `nil` value (see inline docs), we can
make that behavior configurable later if so desired.

I should note that I would've liked to have had the result:

```Go
&foo{name: "one", bar: &foo{name: "one", bar: nil}}}
```

Be presented with a comment:

```Go
&foo{name: "one", bar: &foo{name: "one", bar: nil /* cyclic */}}}
```

But this is in fact *impossible* to do when producing an `ast.Expr`,
and *incredibly difficult* to do within the bounds of the standard
`go/ast` and `go/format` packages. The closest one can get is
producing a [`*go/printer.CommentedNode`](https://golang.org/pkg/go/printer/#CommentedNode)
but this is not a valid `ast.Expr` which means it must be stored
separately, and relative to the other AST nodes. There is a type
`ast.CommentMap`, which tries to help solve this problem but
`go/format` and `go/printer` packages only have a way to supply this
information if you are producing the AST _for an entire file_, not
just a single expression as we are. ¯\_(ツ)_/¯

cc @Andoryuuta

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>