add support for pointers to interfaces
I discovered in 7ffaa053 that we can in fact produce pointers to
unaddressable values (see that commit message for details), so we can
now allow pointers to interfaces which previously were not allowed.

The only annoyance here is that ErrPointerToInterface being removed
is an API-breaking change. I have decided to do this in v1 because:

1. The package itself is really new (<1 week), and most users are
   only making use of the `valast.String()` API which does not expose
   this error.
2. We could keep the type, but it will cause confusion in the future
   and it wouldn't ever be returned anyway, so doing this would
   merely allow us to pretend we didn't make an API-breaking change.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>